Core: Updated to 10.2.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/a4a4d010a0e329d4dbd82c0be5feab1fc06c8834)

+++ Source/Game/Networking/Packets/ChatPackets.cs
@@ -398,7 +398,7 @@ namespace Game.Networking.Packets
 
         public override void Write()
         {
-            _worldPacket.WriteUInt8((byte)Reason);
+            _worldPacket.WriteInt32((int)Reason);
         }
 
         ChatRestrictionType Reason;
